requirements.txt
pathlib2
syjson
IPython
asyncio
python-decouple
google-oauth==1.0.1
orjson >= 3.10,<4
gTTS==2.5.1
gTTS-token==1.1.4
gunicorn==23.0.0
google-api-python-client==2.116.0
fastapi==0.92.0
fasteners==0.19
vectorgrep==1.1.0
uvicorn==0.17.6
websockets==15.0.1
itsdangerous